Bill S1823 version original - Toby Ann Stavisky
BILL NUMBER: S1823 SPONSOR: STAVISKY TITLE OF BILL: An act to amend the executive law, in relation to designation of days of commemoration SUMMARY OF SPECIFIC PROVISIONS: Section 1. Amends subdivision 3 of section 168-a cf the executive law, as amended by chapter 237 of the laws of 2020 law to designate January 13th to be known as Korean American Day. Section 2. Sets forth the effective date.