Bill S1844 version original - Leroy Comrie
BILL NUMBER: S1844 SPONSOR: COMRIE TITLE OF BILL: An act to amend the public authorities law, in relation to adjudication of toll collection violations PURPOSE: To authorize the New York State Thruway Authority and the New York State Bridge Authority to establish an administrative tribunal to adjudicate the liability of owners for violation of toll collection regulations. SUMMARY OF PROVISIONS: Section 1. Adds three new subdivisions, 15, 16 and 17, to section 2985 of the public authorities law.