Bill S1899 version original - Toby Ann Stavisky
BILL NUMBER: S1899 SPONSOR: STAVISKY TITLE OF BILL: An act to amend the executive law, in relation to designating the month of May as Korean American Family Month PURPOSE OR GENERAL IDEA OF BILL: The purpose of the bill is to designate May as Korean American Family Month. SUMMARY OF SPECIFIC PROVISIONS: Section 1.