Bill S1813 version original - Joseph P. Addabbo Jr.
BILL NUMBER: S1813 SPONSOR: ADDABBO TITLE OF BILL: An act to amend the executive law, in relation to revoking community supervision for certain conduct PURPOSE OR GENERAL IDEA OF BILL: To amend certain provisions of the Less of More Act of 2021 to exclude violations that may impact public safety. SUMMARY OF PROVISIONS: Section 1 amends subdivisions 6 and 7 of section 259,of the Executive Law.