Bill S1854 version original - Michelle Hinchey
BILL NUMBER: S1854 SPONSOR: HINCHEY TITLE OF BILL: An act to amend the state law, in relation to designating Lactarius peckii as the official state fungus of the state PURPOSE OR GENERAL IDEA OF BILL:. The purpose of this legislation is to designate a state fungus SUMMARY OF PROVISIONS: Section one amends the state law by adding a new section 93 to designate the Lactarius peckii as the official fungus of the state. Section two states the effective date.