Bill S1927 version original - Joseph A. Griffo
BILL NUMBER: S1927 SPONSOR: GRIFFO TITLE OF BILL: An act to amend the criminal procedure law, in relation to retention of custody of persons found not guilty by reason of mental disease or defect PURPOSE: Directs court to establish a period of commitment for persons found not responsible by reason of mental disease or defect. SUMMARY OF PROVISIONS: Amends Section 330.