Bill S1901A version original - Toby Ann Stavisky
BILL NUMBER: S1901 SPONSOR: STAVISKY TITLE OF BILL: An act to amend the criminal procedure law, in relation to victim state- ments at the sentencing of a defendant for a misdemeanor PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to permit victims of misdemeanor crimes to request to make a statement at the sentencing of the defendant and to afford the court discretion to accept such request. SUMMARY OF SPECIFIC PROVISIONS: Section 1. Provides the short title as "Emma's Law".