Bill S1821 version original - Toby Ann Stavisky
BILL NUMBER: S1821 SPONSOR: STAVISKY TITLE OF BILL: An act to amend the education law, in relation to school districts screening for gifted and talented students PURPOSE OR GENERAL IDEA OF BILL: To ensure that all students are screened for gifted and talented student programming in order to benefit from the resources offered by their schools. SUMMARY OF SPECIFIC PROVISIONS: Section 1.