Bill S1896 version original - Toby Ann Stavisky
BILL NUMBER: S1896 SPONSOR: STAVISKY TITLE OF BILL: An act to amend the penal law, in relation to assault in the third degree PURPOSE OR GENERAL IDEA OF BILL: This bill codifies penalties for those who intentionally injure an oper- ator or passenger of a vehicle for hire or public transportation. This bill would cover taxi drivers and passengers as well as operators and passengers using other means of public transportation. SUMMARY OF SPECIFIC PROVISIONS: Section 1. Amends section 120.