Bill S1958 version original - Joseph P. Addabbo Jr.
BILL NUMBER: S1958 SPONSOR: ADDABBO TITLE OF BILL: An act to establish a task force to study and make recommendations on school safety in the state of New York PURPOSE: To create a New York State Task Force on School Safety, which would consist of professionals in the fields of education, school adminis- tration, mental health, and law enforcement. Four members would be appointed by the Governor, four by the Senate Majority Leader, and four by the Speaker of the Assembly.