Bill S1820 version original - Toby Ann Stavisky
BILL NUMBER: S1820 SPONSOR: STAVISKY TITLE OF BILL: An act to amend the penal law, in relation to including corrections officers in certain assault charges PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to make it a class C felony to assault a corrections officer while such officer is performing their professional duties. SUMMARY OF SPECIFIC PROVISIONS: Section 1. Amends section 120.