Bill S1921 version original - Luis R. Sepúlveda
BILL NUMBER: S1921 SPONSOR: SEPULVEDA TITLE OF BILL: An act to amend the correction law, in relation to establishing emergen- cy release protocols during a time of crisis PURPOSE: To authorize the department of corrections to release certain incarcer- ated individuals in an emergency. SUMMARY OF PROVISIONS: Section 1 adds a new section 807 to the correction law identifying popu- lations eligible to be released in emergency situations.