Bill S1864 version original - Samra Brouk
BILL NUMBER: S1864 SPONSOR: BROUK TITLE OF BILL: An act to amend the tax law, in relation to the donation of a human organ PURPOSE OR GENERAL IDEA OF BILL: Expands the existing tax deduction of $10,000 for human organ donations to make it easier for persons to donate organs without worrying about expenses related to the transplant. SUMMARY OF PROVISIONS: Section 1.