Bill S1961 version original - Joseph P. Addabbo Jr.
BILL NUMBER: S1961 SPONSOR: ADDABBO TITLE OF BILL: An act to amend the education law, in relation to requiring a licensed massage therapist to display his or her license when practicing and the required contents of such license PURPOSE: To require that a license with photo be on display when a massage thera- pist licensed by the State of New York is practicing. SUMMARY OF PROVISIONS: Section 1.