Bill S1974A version original - Jessica Ramos
BILL NUMBER: S1974 SPONSOR: RAMOS TITLE OF BILL: An act to amend the workers' compensation law, in relation to contracted network pharmacy use PURPOSE: To authorize workers' compensation claimants to use any in-state pharma- cy that is registered with the education department. SUMMARY OF PROVISIONS: Sections 1-2 of the bill amend paragraph 5 of subdivision (i) of section 13 of the workers' compensation law, as added by chapter 6 of the laws of 2007.