Bill S1935 version original - Jamaal Bailey
BILL NUMBER: S1935 SPONSOR: BAILEY TITLE OF BILL: An act to amend the family court act, in relation to providing for the confidentiality, sealing, and expungement of juvenile delinquency records; and to repeal section 375.3 of the family court act relating thereto PURPOSE: This bill provides for the confidentiality, sealing, and expungement of juvenile delinquency records; repealer. SUMMARY OF SPECIFIC PROVISIONS: Section 1. The family court act is amended by adding a new section 301.