Bill S1894 version original - Toby Ann Stavisky
BILL NUMBER: S1894 SPONSOR: STAVISKY TITLE OF BILL: An act to amend the penal law, in relation to adding transmission of unwanted lewd or intimate images to the definition of aggravated harass- ment in the second degree PURPOSE OR GENERAL IDEA OF BILL: To penalize intentional harassment of another person via the trans- mission of lewd imagery without the consent of the recipient. SUMMARY OF SPECIFIC PROVISIONS: Section 1. Amends Section 240.