Bill S2065 version original - Toby Ann Stavisky
BILL NUMBER: S2065 SPONSOR: STAVISKY TITLE OF BILL: An act to amend the tax law, in relation to increasing the state excise tax on cigarettes PURPOSE OR GENERAL IDEA OF BILL: This legislation will increase the state excise tax on cigarettes by $1.89. SUMMARY OF SPECIFIC PROVISIONS: Section 1. Amends subdivision 1 of section 471 of the tax law, as amended by section 1 of part D of chapter 134 of the laws of 2010, to change the tax on cigarettes from $4.35 to $6.