Bill S2143A version original - Roxanne J. Persaud
BILL NUMBER: S2143 SPONSOR: PERSAUD TITLE OF BILL: An act to amend the social services law, in relation to the creation of a self-sufficiency standard study PURPOSE OF BILL: The purpose of this bill is to measure the income adequacy in New York State that defines the amount of income necessary to meet the basic needs of families. SUMMARY OF SPECIFIC PROVISIONS: Section one amends the social services law by adding a new § 131-bb.