Bill S2094A version original - Anthony H. Palumbo
BILL NUMBER: S2094 SPONSOR: PALUMBO TITLE OF BILL: An act authorizing the town of Southampton, county of Suffolk, to enact by a local law a homestead exemption and cap on assessment increases for real property tax purposes PURPOSE: Authorizes the town of Southampton, county of Suffolk, to enact a home- stead exemption and cap on assessment increases for real property tax purposes. SUMMARY OF PROVISIONS: Section 1.