Bill S2051 version original - Robert Jackson
BILL NUMBER: S2051 SPONSOR: JACKSON TITLE OF BILL: An act to amend the labor law, in relation to including bonus compen- sation in the definition of wages and forfeit of wages PURPOSE: Labor Law 190(1) does not define bonus compensation as wages even if such bonus compensation is contractually due the employee and is fully earned. The bill corrects this defect in the law.