Bill S2156 version original - Anthony H. Palumbo
BILL NUMBER: S2156 SPONSOR: PALUMBO TITLE OF BILL: An act in relation to the trustees of the Freeholders and the Commonalty of the town of Southampton, county of Suffolk PURPOSE: This legislation clarifies certain powers of the trustees of the Free- holders and the Commonalty of the town of Southampton. SUMMARY OF PROVISIONS: Section 1. Legislative findings. Section 2.