Bill S2123 version original - Joseph P. Addabbo Jr.
BILL NUMBER: S2123 SPONSOR: ADDABBO TITLE OF BILL: An act to amend the state finance law, in relation to making a technical correction to the designation of the animal shelter regulation fund PURPOSE: This legislation will provide comprehensive standards of care for munic- ipal shelters, not for profit humane societies, SPCAs and animal shel- ters and not for profit animal rescues. SUMMARY OF PROVISIONS: Section 1 provides the Legislative Intent.