Bill S2087 version original - Brian Kavanagh
BILL NUMBER: S2087 SPONSOR: KAVANAGH TITLE OF BILL: An act to amend the public authorities law, in relation to enacting the "public benefits for professional facilities act" PURPOSE: To create transparency in the allocation of public benefits to profes- sional sports stadiums, and to make sure stadiums receiving public funds are accessible to the community funding them. SUMMARY OF SPECIFIC PROVISIONS: Section one establishes the act as the "public benefits for professional facilities act.