Bill S2042 version original - James Tedisco
BILL NUMBER: S2042 SPONSOR: TEDISCO TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to the issuance of distinctive license plates to volunteer firefighters PURPOSE: Relates to the issuance of distinctive license plates to volunteer Fire- fighters at no cost. SUMMARY OF PROVISIONS: Subdivision 3 of section 404-b of the vehicle and traffic law, as amended by chapter 277 of the laws of 1991, is amended to read as follows: 3.