Bill S2058 version original - Robert Jackson
BILL NUMBER: S2058 SPONSOR: JACKSON TITLE OF BILL: CONCURRENT RESOLUTION OF THE SENATE AND ASSEMBLY proposing an amendment to section 2 of article 3 of the constitution, in relation to increasing the length of the term of office for members of the legislature PURPOSE: Proposing an amendment to section 2 of article 3 of the constitution, in relation to increasing the length of the term of office for members of the legislature. SUMMARY OF PROVISIONS: Section 1.