Bill S2020 version original - Joseph P. Addabbo Jr.
BILL NUMBER: S2020 SPONSOR: ADDABBO TITLE OF BILL: An act to amend the penal law, in relation to making life imprisonment without parole mandatory for defendants convicted of murder in the first degree and the victim is a police officer PURPOSE: To require a mandatory life sentence without parole for the murder of police officers. SUMMARY OF PROVISIONS: Amends section 60.