Bill S2092 version original - Anthony H. Palumbo
BILL NUMBER: S2092 SPONSOR: PALUMBO TITLE OF BILL: An act to amend the uniform justice court act, in relation to the mone- tary jurisdiction of justice courts PURPOSE: Increased the monetary jurisdiction of the justice courts from $3,000 to $15,000. SUMMARY OF PROVISIONS: This legislation amends language in various sections of the Uniform Justice Court Act to reflect increasing the monetary jurisdiction of justice courts from $3,000 to $15,000.