Bill S2157A version original - Anthony H. Palumbo
BILL NUMBER: S2157 SPONSOR: PALUMBO TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to the restrictions on class DJ and class MJ licenses in the counties of Nassau and Suffolk by permitting such licensees to drive to and from school between the hours of five in the morning to nine o'clock in the evening PURPOSE: To permit holders of Class DJ and Class MJ licenses to drive directly to and from school during the hours of 5:00 a.m. and 9:00 p.m.