Bill S2053 version original - Robert Jackson
BILL NUMBER: S2053 SPONSOR: JACKSON TITLE OF BILL: An act to amend the labor law, in relation to split shifts and minimum wage PURPOSE: To discourage unduly long split shifts or to compensate employees for the same. SUMMARY OF PROVISIONS: Section 1 adds a new subdivision 7 to Section 652 of the Labor Law that provides for the payment of the minimum wage for all time past one hour in which an employee is working a split shift. Section 2 is the effective date.