Bill S2085 version original - Brian Kavanagh
BILL NUMBER: S2085 SPONSOR: KAVANAGH TITLE OF BILL: An act to amend the public health law, in relation to the closure of hospitals or emergency or maternity departments PURPOSE: To allow for greater community input when hospital closures are proposed. SUMMARY OF SPECIFIC PROVISIONS: Section 1 establishes the title as the "Local Input in Community Health- care (the LICH Act).