Bill S2060A version original - Toby Ann Stavisky
BILL NUMBER: S2060 SPONSOR: STAVISKY TITLE OF BILL: An act to amend the education law, in relation to hate crime reporting on college campuses PURPOSE OR GENERAL IDEA OF BILL: This bill amends Article 129-A of the education law to update the poli- cies related to conduct on college campuses regarding hate crimes. SUMMARY OF SPECIFIC PROVISIONS: Section 1.