Bill S2041 version original - James Tedisco
BILL NUMBER: S2041 SPONSOR: TEDISCO TITLE OF BILL: An act to amend the financial services law, in relation to a limitation on financial services assessments and appropriation suballocations PURPOSE: Establishes a limitation on financial services assessments and appropri- ation sub allocations. SUMMARY OF PROVISIONS: Section 1. The financial services law is amended by adding a new section 208 to read as follows: § 208.