Bill S2061 version original - Toby Ann Stavisky
BILL NUMBER: S2061 SPONSOR: STAVISKY TITLE OF BILL: An act to amend the criminal procedure law, the social services law, the family court act and the executive law, in relation to maintaining the confidentiality of immigration status for victims of domestic violence PURPOSE OR GENERAL IDEA OF BILL: To ensure that all women in New York State feel safe to report domestic violence incidents. SUMMARY OF SPECIFIC PROVISIONS: Section 1.