Bill S2075A version original - Joseph P. Addabbo Jr.
BILL NUMBER: S2075 SPONSOR: ADDABBO TITLE OF BILL: An act to amend the education law, in relation to certain school holi- days; and to repeal certain provisions of such law relating thereto PURPOSE OR GENERAL IDEA OF BILL: To allow the designation of Diwali as a school holiday in New York city. SUMMARY OF PROVISIONS: Section 1 is the legislative findings and intent. Section 2 repeals Section 2586 of the education law.