Bill S2106 version original - Jamaal Bailey
BILL NUMBER: S2106 SPONSOR: BAILEY TITLE OF BILL: An act to amend the administrative code of the city of New York, in relation to prohibiting wind turbines from being located within 1500 feet of residential areas and/or commercial shopping districts PURPOSE: The Purpose of the bill is to prohibit the construction and placement of wind turbines near populated areas. SUMMARY OF SPECIFIC PROVISIONS: The new section 27-521.