Bill S2109 version original - Anthony H. Palumbo
BILL NUMBER: S2109 SPONSOR: PALUMBO TITLE OF BILL: An act to amend the county law, in relation to establishing a procedure for the creation of new counties PURPOSE: Establishes a procedure for the creation of new counties. SUMMARY OF PROVISIONS: Section 1. Legislative Findings Section 2. The county law is amended by adding a new article 1-a, "Creation of New Counties". Section 3. Severability Clause Section 4.