Bill S1985 version original - Robert Jackson
BILL NUMBER: S1985 SPONSOR: JACKSON TITLE OF BILL: An act to amend the transportation law, in relation to enacting the community gridlock prevention act PURPOSE: The purpose of this bill is to include traffic impacts on land use actions and proposals. This bill would prevent land use actions if the findings of the impact study would cause systematic traffic gridlock in a neighborhood or planning area. SUMMARY OF PROVISIONS: Section 1 is the short title.