Bill S2115 version original - Simcha Felder
BILL NUMBER: S2115 SPONSOR: FELDER TITLE OF BILL: An act to enact the "save our stores act" SUMMARY OF PROVISIONS: 1. Title "Save our Stores Act." 2. Section 1, Paragraph (a) In a city with a population of one million or more, any fine assessed against a not-for-profit organization, place of worship, school or small business by a state or local agency for a violation during the COVID-19 covered period shall be deemed void and waived unless it falls under paragraph b Paragraph (b) lists severa...