Bill S2029 version original - Dan Stec
BILL NUMBER: S2029 SPONSOR: STEC TITLE OF BILL: An act relating to granting a waiver to Adirondack Helping Hands for the 2016--2017 school year PURPOSE: The bill ratifies a waiver for $52,000 in spending by Adirondack Helping Hands Inc. in the 2016-17 school year relating to its 9160 program, and prohibits the state from imposing penalties in subsequent years. SUMMARY OF PROVISIONS: Section 1 - Grants a waiver for $52,000 relating to a special class in an integrated setting.