Bill S2100A version original - Pete Harckham
BILL NUMBER: S2100 SPONSOR: HARCKHAM TITLE OF BILL: An act to amend the election law, in relation to establishing a code of ethics for all board of election employees PURPOSE:: Establishes a code of ethics for all board of elections employees SUMMARY OF PROVISIONS:: Section 1 amends the election law by adding section 3-201 creating the board of elections; employee code of ethics. The employee code of ethics includes rules with respect to conflicts of interest and creates a set of standards.