Bill S1994 version original - Rachel May
BILL NUMBER: S1994 SPONSOR: MAY TITLE OF BILL: An act to amend the legislative law and the election law, in relation to prohibiting certain functions whenever the legislature is in session PURPOSE OR GENERAL IDEA OF BILL: The purpose of this legislation is to prohibit fundraising events within 15 miles of the Capitol when the legislature is in Session. SUMMARY OF SPECIFIC PROVISIONS: Section 1.