Bill S2068 version original - Toby Ann Stavisky
BILL NUMBER: S2068 SPONSOR: STAVISKY TITLE OF BILL: An act to amend the education law, in relation to allowing for a limited permit to be obtained to practice as a chiropractor PURPOSE OR GENERAL IDEA OF BILL: To create a limited license permit so that qualified graduates of a chiropractic school to practice under licensed supervision until such time as they may take in-person examinations for licensure. SUMMARY OF SPECIFIC PROVISIONS: Section 1.