Bill S2076 version original - Joseph P. Addabbo Jr.
BILL NUMBER: S2076 SPONSOR: ADDABBO TITLE OF BILL: An act to amend the racing, pari-mutuel wagering and breeding law, in relation to certain payments to the horsemen's organization PURPOSE: The purpose of this legislation is to extend the authority of the New York Thoroghbred Horsmen's Association to receive an additional 1.0% of purse funds for the support of additional medical and mental health services for backstretch employees. SUMMARY OF PROVISIONS: Section 1.