Bill S2136 version original - John Liu
BILL NUMBER: S2136 SPONSOR: LIU TITLE OF BILL: An act to amend the education law, in relation to requiring the filing and posting of required ancillary agreements on school district and charter school websites PURPOSE OR GENERAL IDEA OF BILL: This bill requires filing and posting of required ancillary agreements on the school and school district websites. SUMMARY OF SPECIFIC PROVISIONS: Section 1 amends Section 2116-a by adding a new subdivision 5 S 2.