Bill S2002 version original - Rachel May
BILL NUMBER: S2002 SPONSOR: MAY TITLE OF BILL: An act to amend the tax law, in relation to imposing a specific tax on all corporate stock buybacks of issued shares PURPOSE: To impose a tax on corporate stock buybacks of issued shares. SUMMARY OF PROVISION: Section 1 - amends subdivisions 1 and 2 of section 270 of the tax law, to impose a tax on the reacquisition of stock by a corporation even if such corporation subsequently cancels such stock.