Bill S2131 version original - Liz Krueger
BILL NUMBER: S2131 SPONSOR: KRUEGER TITLE OF BILL: An act to amend the penal law and the criminal procedure law, in relation to bribery and abuse of public trust PURPOSE: To strengthen the penalties for public officials who participate in acts of bribery, attempted bribery and other crimes that are committed through their position as a public official. SUMMARY OF PROVISIONS: Sections 200.00 200.03, 200.