Bill S2150 version original - Joseph P. Addabbo Jr.
BILL NUMBER: S2150 SPONSOR: ADDABBO TITLE OF BILL: An act to amend the environmental conservation law, in relation to prohibiting the filling of borrow pits in Jamaica Bay PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to extend permanently, requirements to continue protection of Jamaica Bay. This will strengthen protection of the Jamaica Bay waters making it safer for current and all future gener- ations.