Bill S2175 version original - Timothy M. Kennedy
BILL NUMBER: S2175REVISED 1/19/23 SPONSOR: KENNEDY TITLE OF BILL: An act to amend the workers' compensation law, in relation to providing for paid family leave following a stillbirth PURPOSE OR GENERAL IDEA OF BILL: This bill provides for paid family leave following the stillbirth of a child. SUMMARY OF SPECIFIC PROVISIONS: Section one amends the workers' compensation law to include "recovery after a stillbirth" as a qualifying condition for family leave.