Bill S2240 version original - Alexis Weik
BILL NUMBER: S2240 SPONSOR: WEIK TITLE OF BILL: An act to amend the executive law, in relation to establishing the defund municipalities that defund the police act PURPOSE: To withhold state funding in proportion to the amount a municipality defends its police department. SUMMARY OF PROVISIONS: Sections 1 & 2-Sections 1 & 2 create the defund municipalities that defund the police act.