Bill S2184 version original - Thomas F. O'Mara
BILL NUMBER: S2184 SPONSOR: O'MARA TITLE OF BILL: An act to amend the tax law, in relation to establishing a home heating tax credit; and providing for the repeal of such provisions upon expira- tion thereof PURPOSE OR GENERAL IDEA OF BILL: This bill would provide a home heating tax credit. SUMMARY OF SPECIFIC PROVISIONS: This bill would provide a home heating tax credit for individuals with incomes of up to $125,000 and married couples with income up to $250,000.