Bill S2266 version original - Liz Krueger
BILL NUMBER: S2266 SPONSOR: KRUEGER TITLE OF BILL: An act to amend the penal law, in relation to creating the offense of defrauding the government in the first degree and expanding the defi- nition of conspiracy in the fourth degree; and to amend the tax law and the administrative code of the city of New York, in relation to expand- ing the definition of tax fraud acts PURPOSE: 1. Creates the offense of defrauding the government in the first degree. 2.