Bill S2284 version original - Cordell Cleare
BILL NUMBER: S2284 SPONSOR: CLEARE TITLE OF BILL: An act to amend the general business law, in relation to manufacturers' reporting requirements for the sale or distribution of firearms PURPOSE OR GENERAL IDEA OF BILL: To require manufacturers of firearms sold in New York to report to the Department of Criminal Justice Services the number of firearms they manufacture and sell on an annual basis. SUMMARY OF PROVISIONS: The general business law is amended by adding a new article 39-DDDD. S 898-a.