Bill S2176 version original - Thomas F. O'Mara
BILL NUMBER: S2176 SPONSOR: O'MARA TITLE OF BILL: An act to amend the penal law, in relation to requiring the division of state police to send renewal notices to holders of a license to carry or possess a pistol or revolver before expiration of such license PURPOSE OR GENERAL IDEA OF BILL: Requires the division of state police to send renewal notices to holders of a. license to carry or possess a pistol or revolver at least ninety days before expiration of such license.