Bill S2288 version original - Pamela Helming
BILL NUMBER: S2288 SPONSOR: HELMING TITLE OF BILL: An act to amend the real property law, in relation to the rights of tenants to display the flag of the United States PURPOSE: To allow those who rent property to have the privilege to display the American flag. SUMMARY OF PROVISIONS: Section 1: Amends the Real Property Law by adding a new section 223-c. Section 2: Provides the effective date.