Bill S2235 version original - Shelley Mayer
BILL NUMBER: S2235 SPONSOR: MAYER TITLE OF BILL: An act to amend the civil rights law, in relation to the amount of damages recoverable in a cause of action for unlawful interference with protected rights PURPOSE: This is a chapter amendment to L.2022, ch.2018, to modify the types of damages available in a claim brought pursuant to the underlying chapter. SUMMARY OF PROVISIONS: This chapter amendment strikes language from the underlying chapter that authorized the award of punitive damages.