Bill S2234 version original - Brad Hoylman-Sigal
BILL NUMBER: S2234 SPONSOR: HOYLMAN-SIGAL TITLE OF BILL: An act to amend the domestic relations law, in relation to designating lay individuals to solemnize marriages; and to repeal certain provisions of the executive law related thereto PURPOSE: This is a chapter amendment to L.2022, ch.798, to modify the process by which a layperson applies for a one-day marriage officiant license. SUMMARY OF PROVISIONS: This legislation amends multiple provisions of L2022, ch.798.