Bill S2163B version original - Joseph P. Addabbo Jr.
BILL NUMBER: S2163 SPONSOR: ADDABBO TITLE OF BILL: An act to amend the agriculture and markets law, in relation to prohib- iting the slaughtering of horses for human consumption PURPOSE: To amend the agriculture and markets law by adding a new section 382, to prohibit the slaughtering of horses for human consumption. SUMMARY OF PROVISIONS: The agriculture and markets law shall be amended by adding a new section 382. This section will make it unlawful to slaughter a horse for human consumption.