Bill S2174 version original - James Tedisco
BILL NUMBER: S2174 SPONSOR: TEDISCO TITLE OF BILL: An act to amend the tax law, in relation to establishing limitations on excise and sales taxes and petroleum business taxes on diesel motor fuel and motor fuel and to authorize cities having a population of one million or more and counties to adopt local laws limiting taxes on diesel motor fuel and motor fuel PURPOSE: Establishes limitations on taxes on diesel motor fuel and motor fuel. SUMMARY OF PROVISIONS: § 1.