Bill S2217 version original - Thomas F. O'Mara
BILL NUMBER: S2217 SPONSOR: O'MARA TITLE OF BILL: An act to amend the penal law, in relation to the possession and sale of methamphetamine PURPOSE OF BILL: To amend the penal law, so that the penalties for the possession and or sale of methamphetamine are similar to that of the penalties for possession and/or sale of heroin and cocaine. SUMMARY OF PROVISIONS: Amends the penal law by adding a new subdivision 8 to section 220.