Bill S2322 version original - Jamaal Bailey
BILL NUMBER: S2322 SPONSOR: BAILEY TITLE OF BILL: An act to amend the public officers law, in relation to the disclosure of law enforcement officer's disciplinary records PURPOSE OR GENERAL IDEA OF BILL: The purpose of this legislation is to clarify the full scope of § 50-a's repeal. SUMMARY OF SPECIFIC PROVISIONS: Section 1 amends section 87 of the public officers' law by adding a new subdivision, 4-c, which further clarifies situations in which a request for records cannot be denied.