Bill S2243 version original - Alexis Weik
BILL NUMBER: S2243 SPONSOR: WEIK TITLE OF BILL: An act relating to publishing data on the effects of bail reform PURPOSE: To require OCA and DCJS to collect and provide data as it relates to individuals charged with crimes that no longer qualify for bail. SUMMARY OF PROVISIONS: Section 1 is the legislative intent.