Bill S2241 version original - Alexis Weik
BILL NUMBER: S2241 SPONSOR: WEIK TITLE OF BILL: An act to amend the executive law, in relation to designating March twenty-first as "Down Syndrome Awareness Day" PURPOSE: To commemorate March 21 as Down Syndrome Awareness Day in the State of New York. SUMMARY OF PROVISIONS: Section 1 amends section 168-a(3) of the executive law to amend March 21 as Down Syndrome Awareness Day to the list of commemoration days in the State of New York. Section 2 sets forth an immediate effective date.