Bill S2323A version original - Jamaal Bailey
BILL NUMBER: S2323 SPONSOR: BAILEY TITLE OF BILL: An act to amend the public health law, in relation to providing for the New York state of health to help veterans ascertain their eligibility for healthcare through the department of veterans affairs PURPOSE: The bill provides veterans with a local intake coordinator when they enroll in Medicaid through the United States department of veteran affairs. SUMMARY OF PROVISIONS: Section 1.