Bill S2179 version original - Thomas F. O'Mara
BILL NUMBER: S2179 SPONSOR: O'MARA TITLE OF BILL: An act to amend the penal law, in relation to including fires caused by the manufacture of a controlled substance within the crime of arson in the first degree PURPOSE OR GENERAL IDEA OF BILL: To amend the Arson statute to include individuals who illegally manufac- ture a controlled substance that results in an explosion or fire damage to a building or person. SUMMARY OF SPECIFIC PROVISIONS: Section 1. amends subdivision 2 of section 150.