Bill S2212 version original - Thomas F. O'Mara
BILL NUMBER: S2212 SPONSOR: O'MARA TITLE OF BILL: An act to amend the environmental conservation law, in relation to the transportation of liquefied natural and petroleum gas, and to repeal certain provisions of such law relating thereto PURPOSE: To amend the environmental conservation law to repeal paragraphs b and c of subdivision 2 and paragraph c of subdivision 3 of section 23-1709 of the environmental conservation law. SUMMARY OF PROVISIONS: Section 1.