Bill S2203 version original - Robert Ortt
BILL NUMBER: S2203 SPONSOR: ORTT TITLE OF BILL: An act to amend the veterans' services law, in relation to permitting private employers to adopt a veterans' preference employment policy PURPOSE: To allow private companies to adopt a veterans' preference for hiring, promoting, or retaining a veteran. SUMMARY OF PROVISIONS: Section 1 Allows a private employer to adopt a veterans' preference for hiring, promoting, or retaining a veteran.