Bill S2327 version original - Jamaal Bailey
BILL NUMBER: S2327 SPONSOR: BAILEY TITLE OF BILL: An act to amend the tax law, in relation to creating a tax deduction for certain capital gains PURPOSE OR GENERAL IDEA OF BILL: This bill creates a complete capital gains tax exemption for sale of stock to an employee-owned enterprise. SUMMARY OF SPECIFIC PROVISIONS: Section 1 amends the tax law to provide complete capital gains tax exemption for sale of stock to an employee-owned enterprise. Section 2 is the effective date.