Bill S2236 version original - Michelle Hinchey
BILL NUMBER: S2236 SPONSOR: HINCHEY TITLE OF BILL: An act to amend the agriculture and markets law, in relation to estab- lishing a one-stop farming hotline PURPOSE: To provide farmers with access to essential and valuable information on all aspects of farming, namely, tax credit programs, beginning farmer loans, and sustainable practices. SUMMARY OF PROVISIONS: Section 1 adds a new provision at section 216 of the Agriculture and.