Bill S2173 version original - John Liu
BILL NUMBER: S2173 SPONSOR: LIU TITLE OF BILL: An act to amend the education law, in relation to charter schools in cities with a population of one million or more PURPOSE OR GENERAL IDEA OF BILL: Allows the New York State Comptroller to conduct audits of New York City charter schools. SUMMARY OF PROVISIONS: Section one provides that charter schools located in the city of New York are subject to financial audits by the comptroller of the State of New York.