Bill S2185 version original - Thomas F. O'Mara
BILL NUMBER: S2185 SPONSOR: O'MARA TITLE OF BILL: An act to authorize a correction to an error in the tax assessment rolls of the Bath Central School District; and making an appropriation there- for PURPOSE: To remedy a clerical error that occurred during the Town of Bath reeval- uation project. SUMMARY OF SPECIFIC PROVISIONS: Section 1 of this bill authorizes the assessor of the town of Bath to make the appropriate correction to the 2018 tax assessment roll.