Bill S2251A version original - Thomas F. O'Mara
BILL NUMBER: S2251 SPONSOR: O'MARA TITLE OF BILL: An act to amend the tax law, in relation to authorizing an occupancy tax in the village of Bath, in Steuben county; and providing for the repeal of such provisions upon expiration thereof PURPOSE OR GENERAL IDEA OF BILL: To authorize the Village of Bath (Steuben County) to impose a hotel/motel occupancy tax up to two percent. SUMMARY OF PROVISIONS: Section 1 of the bill amends the tax law to add a new section 1202-z-4.