Bill S2214 version original - Thomas F. O'Mara
BILL NUMBER: S2214 SPONSOR: O'MARA TITLE OF BILL: An act to amend the penal law, in relation to penalties for the crimes of criminal possession of methamphetamine manufacturing material and unlawful manufacture of methamphetamine PURPOSE: To amend the penal law, increasing the penalties for the criminal possession of methamphetamine manufacturing material and unlawful manu- facture of methamphetamine. SUMMARY OF PROVISIONS: Section 1. 2.