Bill S2187 version original - Jamaal Bailey
BILL NUMBER: S2187 SPONSOR: BAILEY TITLE OF BILL: An act to amend the state finance law, in relation to establishing the New York baby opportunity fund PURPOSE: Each newborn in New York State households earning an income of $125,000 or less will receive $1,000, which will be put into a savings account. The account will be accessible on the child's 18th birthday, and can only be used for education, the purchase of a home, or starting a busi- ness. SUMMARY OF SPECIFIC PROVISIONS: Section.