Bill S2254 version original - Anthony H. Palumbo
BILL NUMBER: S2254 SPONSOR: PALUMBO TITLE OF BILL: An act to amend the education law, in relation to eligibility for an apportionment; and making an appropriation therefor PURPOSE: This legislation amends the education law to provide an apportionment to each school district participating in a shared superintendent program and creating an appropriation for the State Education Department to carry out provisions of this section. SUMMARY OF PROVISIONS: Section 1.