Bill S2320 version original - Jamaal Bailey
BILL NUMBER: S2320 SPONSOR: BAILEY TITLE OF BILL: An act to amend the criminal procedure law, in relation to the timely scheduling of criminal trials PURPOSE OR GENERAL IDEA OF BILL: This bill amends section 30.30 of the Criminal Procedure Law. SUMMARY OF SPECIFIC PROVISIONS: Paragraph (d) of subdivisions 1 and 2 ensures that traffic infractions charged alongside a Misdemeanor in New York City, are covered by the same provisions of 30.30 as other traffic infractions in the state.