Bill S2286 version original - Cordell Cleare
BILL NUMBER: S2286 SPONSOR: CLEARE TITLE OF BILL: An act to amend the insurance law, in relation to enacting the "Give Kids a Chance - Carter's Law" mandating health insurance coverage for congenital anomalies PURPOSE: To require coverage for medical and dental treatments as well as rehabi- litative therapies (physical, speech, occupational, etc.) to address conditions or illness related to or developed as a result of a congen- ital anomaly. SUMMARY OF PROVISIONS: Section 1.