Bill S2324 version original - Jamaal Bailey
BILL NUMBER: S2324 SPONSOR: BAILEY TITLE OF BILL: An act to amend the general business law, in relation to disclosing to a parent the personal information and content about a minor collected by an operator of an internet platform PURPOSE OR GENERAL IDEA OF BILL: Relates to disclosing to a parent the person information and content about a minor collected by an operator of an internet platform. SUMMARY OF SPECIFIC PROVISIONS: Section 1 provides important definitions.