Bill S2210 version original - Thomas F. O'Mara
BILL NUMBER: S2210 SPONSOR: O'MARA TITLE OF BILL: An act to amend the environmental conservation law, in relation to including the mourning dove within the definition of "migratory game birds" and to allow for their taking PURPOSE OR GENERAL IDEA OF BILL: Relates to including the mourning dove within the definition of "migra- tory game birds" and to allow for their taking. SUMMARY OF SPECIFIC PROVISIONS: Section 1.