Bill S2263 version original - James Sanders Jr.
BILL NUMBER: S2263 SPONSOR: SANDERS TITLE OF BILL: An act to amend the real property actions and proceedings law, in relation to prohibiting the eviction of tenants for the certified medical use of medical marihuana PURPOSE: To protect tenants from eviction as a result of their certified use of medical marihuana. SUMMARY OF PROVISIONS: Section 1 adds a new section 744-a to the real property law.