Bill S2352 version original - Brian Kavanagh
BILL NUMBER: S2352 SPONSOR: KAVANAGH TITLE OF BILL: An act to amend the real property law, in relation to requiring the secretary of state to require that real estate brokers and salespersons compile and disclose certain demographic information PURPOSE: Requires the Secretary of State to collect from real estate brokers and salespersons the demographic data of their clients. Requires the Secre- tary of State to maintain a database of and issue an annual report of this information.