Bill S2464B version original - Roxanne J. Persaud
BILL NUMBER: S2464 SPONSOR: PERSAUD TITLE OF BILL: An act to amend the general business law, in relation to requiring car washes to disclose when promotions expire PURPOSE OR GENERAL IDEA OF BILL: The purpose of the bill is to increase the transparency between a carwash customer and a carwash business regarding promotion expiration dates. SUMMARY OF PROVISIONS: Section 1. Provides for the definition of "carwash" and carwash promotion disclosure. Section 2. Provides for effective date.