Bill S2345 version original - Zellnor Myrie
BILL NUMBER: S2345 Revised 4/13/23 SPONSOR: MYRIE TITLE OF BILL: An act to amend the correction law, in relation to enacting "the prison minimum wage act" PURPOSE: To end the last vestiges of slavery and embrace the spirit and the prom- ise of the Thirteenth Amendment of the United States Constitution by providing a minimum wage of $3.00 an hour to inmates. SUMMARY OF PROVISIONS: Section 1 provides the title of the bill, "Prison Minimum Wage Act.