Bill S2341 version original - Gustavo Rivera
BILL NUMBER: S2341 SPONSOR: RIVERA TITLE OF BILL: An act to amend the education law, in relation to the use and sale of tobacco at state-operated institutions of the state university of New York PURPOSE: Relates to the use and sale of tobacco at state-operated institutions of the state university of New York. SUMMARY OF PROVISIONS: Section 1 provides legislative intent.